requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
python-socketio==5.10.0
mysql-connector-python==8.0.33
eventlet==0.33.3 # Required for gunicorn with Flask-SocketIO
python-dotenv==1.0.0 # For loading .env file locally
gunicorn==20.1.0 # For production deployment